requirements.txt
flask==3.0.3
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0
